Create ps.as
this is the file i've tried to run. this is from our maman. (after the macro) we need to test with the macro also..

PS: the index addressing in his maman is without the register ( LOOP[4], r6) 
in our maman its with! ( LOOP[r4], r6)
here: rows 11, 14

we need to verify '-' (minus)